Bootloader/V2/commandline/: Firmware/Trunk/usb/: Configuration: bugs solved, EEPROM programming and reading as well as plain-HEX-to-I8HEX conversion working (not stress-tested)
git-svn-id: https://repository.combo-convention.de/dualstrike@108 a2e5ad42-ccce-4154-a8df-8a7cfdfd1d97

+++ Configuration/java/dualstrike/configuration/HexFilesUtility.java
@@ -118,7 +118,7 @@ public class HexFilesUtility {
 				for(int j = 0; j < line.length(); j += 2) {
 					byte b;
 					
-					b = Byte.parseByte(line.substring(j, j + 2), 16);
+					b = (byte)Integer.parseInt(line.substring(j, j + 2), 16);
 					bytes.add(b);
 				}
 			}
